move WanFile data read to model
